Update CMakeLists.txt
The package failed to be built after cloning it in my workspace. This is because the project names in the package.xml and CMakeLists.txt files are not the same. The project name in package.xml is (cmu_pocketsphinx) while in the CMakeLists.txt file the name is (pocketsphinx). Both names must be identical.